+++ sdks/go/pkg/beam/core/funcx/fn.go
@@ -80,6 +80,8 @@ const (
 	FnPane FnParamKind = 0x400
 	// FnBundleFinalization indicates a function input parameter that implements typex.BundleFinalization.
 	FnBundleFinalization FnParamKind = 0x800
+	// FnWatermarkEstimator indicates a function input parameter that implements sdf.WatermarkEstimator
+	FnWatermarkEstimator FnParamKind = 0x1000
 )
 
 func (k FnParamKind) String() string {
@@ -108,6 +110,8 @@ func (k FnParamKind) String() string {
 		return "Pane"
 	case FnBundleFinalization:
 		return "BundleFinalization"
+	case FnWatermarkEstimator:
+		return "WatermarkEstimator"
 	default:
 		return fmt.Sprintf("%v", int(k))
 	}
@@ -285,6 +289,17 @@ func (u *Fn) BundleFinalization() (pos int, exists bool) {
 	return -1, false
 }
 
+// WatermarkEstimator returns (index, true) iff the function expects a
+// parameter that implements sdf.WatermarkEstimator.
+func (u *Fn) WatermarkEstimator() (pos int, exists bool) {
+	for i, p := range u.Param {
+		if p.Kind == FnWatermarkEstimator {
+			return i, true
+		}
+	}
+	return -1, false
+}
+
 // Error returns (index, true) iff the function returns an error.
 func (u *Fn) Error() (pos int, exists bool) {
 	for i, p := range u.Ret {
@@ -363,6 +378,8 @@ func New(fn reflectx.Func) (*Fn, error) {
 			kind = FnType
 		case t.Implements(reflect.TypeOf((*sdf.RTracker)(nil)).Elem()):
 			kind = FnRTracker
+		case t.Implements(reflect.TypeOf((*sdf.WatermarkEstimator)(nil)).Elem()):
+			kind = FnWatermarkEstimator
 		case typex.IsContainer(t), typex.IsConcrete(t), typex.IsUniversal(t):
 			kind = FnValue
 		case IsEmit(t):
@@ -447,7 +464,7 @@ func SubReturns(list []ReturnParam, indices ...int) []ReturnParam {
 }
 
 // The order of present parameters and return values must be as follows:
-// func(FnContext?, FnPane?, FnWindow?, FnEventTime?, FnType?, FnBundleFinalization?, FnRTracker?, (FnValue, SideInput*)?, FnEmit*) (RetEventTime?, RetOutput?, RetError?)
+// func(FnContext?, FnPane?, FnWindow?, FnEventTime?, FnWatermarkEstimator?, FnType?, FnBundleFinalization?, FnRTracker?, (FnValue, SideInput*)?, FnEmit*) (RetEventTime?, RetOutput?, RetError?)
 //     where ? indicates 0 or 1, and * indicates any number.
 //     and  a SideInput is one of FnValue or FnIter or FnReIter
 // Note: Fns with inputs must have at least one FnValue as the main input.
@@ -471,14 +488,15 @@ func validateOrder(u *Fn) error {
 }
 
 var (
-	errContextParam                 = errors.New("may only have a single context.Context parameter and it must be the first parameter")
-	errPaneParamPrecedence          = errors.New("may only have a single PaneInfo parameter and it must precede the WindowParam, EventTime and main input parameter")
-	errWindowParamPrecedence        = errors.New("may only have a single Window parameter and it must precede the EventTime and main input parameter")
-	errEventTimeParamPrecedence     = errors.New("may only have a single beam.EventTime parameter and it must precede the main input parameter")
-	errReflectTypePrecedence        = errors.New("may only have a single reflect.Type parameter and it must precede the main input parameter")
-	errRTrackerPrecedence           = errors.New("may only have a single sdf.RTracker parameter and it must precede the main input parameter")
-	errBundleFinalizationPrecedence = errors.New("may only have a single BundleFinalization parameter and it must precede the main input parameter")
-	errInputPrecedence              = errors.New("inputs parameters must precede emit function parameters")
+	errContextParam                      = errors.New("may only have a single context.Context parameter and it must be the first parameter")
+	errPaneParamPrecedence               = errors.New("may only have a single PaneInfo parameter and it must precede the WindowParam, EventTime and main input parameter")
+	errWindowParamPrecedence             = errors.New("may only have a single Window parameter and it must precede the EventTime and main input parameter")
+	errEventTimeParamPrecedence          = errors.New("may only have a single beam.EventTime parameter and it must precede the main input parameter")
+	errWatermarkEstimatorParamPrecedence = errors.New("may only have a single sdf.WatermarkEstimator parameter and it must precede the main input parameter")
+	errReflectTypePrecedence             = errors.New("may only have a single reflect.Type parameter and it must precede the main input parameter")
+	errRTrackerPrecedence                = errors.New("may only have a single sdf.RTracker parameter and it must precede the main input parameter")
+	errBundleFinalizationPrecedence      = errors.New("may only have a single BundleFinalization parameter and it must precede the main input parameter")
+	errInputPrecedence                   = errors.New("inputs parameters must precede emit function parameters")
 )
 
 type paramState int
@@ -489,6 +507,7 @@ const (
 	psPane
 	psWindow
 	psEventTime
+	psWatermarkEstimator
 	psType
 	psInput
 	psOutput
@@ -508,6 +527,8 @@ func nextParamState(cur paramState, transition FnParamKind) (paramState, error) 
 			return psWindow, nil
 		case FnEventTime:
 			return psEventTime, nil
+		case FnWatermarkEstimator:
+			return psWatermarkEstimator, nil
 		case FnType:
 			return psType, nil
 		case FnBundleFinalization:
@@ -523,6 +544,8 @@ func nextParamState(cur paramState, transition FnParamKind) (paramState, error) 
 			return psWindow, nil
 		case FnEventTime:
 			return psEventTime, nil
+		case FnWatermarkEstimator:
+			return psWatermarkEstimator, nil
 		case FnType:
 			return psType, nil
 		case FnBundleFinalization:
@@ -536,6 +559,8 @@ func nextParamState(cur paramState, transition FnParamKind) (paramState, error) 
 			return psWindow, nil
 		case FnEventTime:
 			return psEventTime, nil
+		case FnWatermarkEstimator:
+			return psWatermarkEstimator, nil
 		case FnType:
 			return psType, nil
 		case FnBundleFinalization:
@@ -547,6 +572,8 @@ func nextParamState(cur paramState, transition FnParamKind) (paramState, error) 
 		switch transition {
 		case FnEventTime:
 			return psEventTime, nil
+		case FnWatermarkEstimator:
+			return psWatermarkEstimator, nil
 		case FnType:
 			return psType, nil
 		case FnBundleFinalization:
@@ -555,6 +582,17 @@ func nextParamState(cur paramState, transition FnParamKind) (paramState, error) 
 			return psRTracker, nil
 		}
 	case psEventTime:
+		switch transition {
+		case FnWatermarkEstimator:
+			return psWatermarkEstimator, nil
+		case FnType:
+			return psType, nil
+		case FnBundleFinalization:
+			return psBundleFinalization, nil
+		case FnRTracker:
+			return psRTracker, nil
+		}
+	case psWatermarkEstimator:
 		switch transition {
 		case FnType:
 			return psType, nil
@@ -598,6 +636,8 @@ func nextParamState(cur paramState, transition FnParamKind) (paramState, error) 
 		return -1, errWindowParamPrecedence
 	case FnEventTime:
 		return -1, errEventTimeParamPrecedence
+	case FnWatermarkEstimator:
+		return -1, errWatermarkEstimatorParamPrecedence
 	case FnType:
 		return -1, errReflectTypePrecedence
 	case FnBundleFinalization: